Xcode maintenance. Get the sample app building again.

+++ JelloSwiftTests/PointMassTests.swift
@@ -11,6 +11,7 @@
     #elseif os(OSX)
     import Cocoa
 #endif
+import CoreGraphics
 import Jello
 import XCTest
 
@@ -18,7 +19,7 @@ class PointMassTests: XCTestCase
 {
     func testExample()
     {
-        var p = PointMass(mass: 0.2, position: Vector2(0, 0));
+        var p = PointMass(mass: CGFloat(0.2), position: Vector2(0, 0));
         p.force += Vector2(1, 2);
         p.force += Vector2(1, 2);
         p.force += Vector2(1, 2);
